Refactor module loading to make it easier to read
