+++ src/init/ad-free.ts
@@ -1,0 +1,29 @@
+import { bootCommercial, type Modules } from '../lib/init-utils';
+import { adFreeSlotRemove } from './consented/ad-free-slot-remove';
+import { init as initComscore } from './consented/comscore';
+import { init as initIpsosMori } from './consented/ipsos-mori';
+import { removeDisabledSlots as closeDisabledSlots } from './consented/remove-slots';
+import { initTeadsCookieless } from './consented/teads-cookieless';
+import { init as initTrackGpcSignal } from './consented/track-gpc-signal';
+import { init as initTrackScrollDepth } from './consented/track-scroll-depth';
+
+// modules not related to ad loading
+const commercialModules: Modules = [
+	['cm-adFreeSlotRemoveFronts', adFreeSlotRemove],
+	['cm-closeDisabledSlots', closeDisabledSlots],
+	['cm-comscore', initComscore],
+	['cm-ipsosmori', initIpsosMori],
+	['cm-teadsCookieless', initTeadsCookieless],
+	['cm-trackScrollDepth', initTrackScrollDepth],
+	['cm-trackGpcSignal', initTrackGpcSignal],
+];
+
+const bootCommercialWhenReady = () => {
+	if (!!window.guardian.mustardCut || !!window.guardian.polyfilled) {
+		void bootCommercial(commercialModules);
+	} else {
+		window.guardian.queue.push(() => bootCommercial(commercialModules));
+	}
+};
+
+export { bootCommercialWhenReady };
