+++ src/lib/init-utils.ts
@@ -1,0 +1,96 @@
+import { log } from '@guardian/libs';
+import { adSlotIdPrefix } from '../lib/dfp/dfp-env-globals';
+import { reportError } from '../lib/error/report-error';
+import { catchErrorsAndReport } from './error/robust';
+import { EventTimer } from './event-timer';
+
+type Modules = Array<[`${string}-${string}`, () => Promise<unknown>]>;
+
+const tags: Record<string, string> = {
+	bundle: 'standalone',
+};
+
+const loadModules = (modules: Modules) => {
+	const modulePromises: Array<Promise<unknown>> = [];
+
+	modules.forEach((module) => {
+		const [moduleName, moduleInit] = module;
+
+		catchErrorsAndReport(
+			[
+				[
+					moduleName,
+					function pushAfterComplete(): void {
+						const result = moduleInit();
+						modulePromises.push(result);
+					},
+				],
+			],
+			tags,
+		);
+	});
+
+	return Promise.allSettled(modulePromises).then(() => {
+		EventTimer.get().mark('commercialModulesLoaded');
+	});
+};
+
+const recordCommercialMetrics = () => {
+	const eventTimer = EventTimer.get();
+	eventTimer.mark('commercialBootEnd');
+	eventTimer.mark('commercialModulesLoaded');
+	// record the number of ad slots on the page
+	const adSlotsTotal = document.querySelectorAll(
+		`[id^="${adSlotIdPrefix}"]`,
+	).length;
+	eventTimer.setProperty('adSlotsTotal', adSlotsTotal);
+
+	// and the number of inline ad slots
+	const adSlotsInline = document.querySelectorAll(
+		`[id^="${adSlotIdPrefix}inline"]`,
+	).length;
+	eventTimer.setProperty('adSlotsInline', adSlotsInline);
+};
+
+const bootCommercial = async (modules: Modules): Promise<void> => {
+	log('commercial', '📦 standalone.commercial.ts', __webpack_public_path__);
+	if (process.env.COMMIT_SHA) {
+		log(
+			'commercial',
+			`@guardian/commercial commit https://github.com/guardian/commercial/blob/${process.env.COMMIT_SHA}`,
+		);
+	}
+
+	// Init Commercial event timers
+	EventTimer.init();
+
+	catchErrorsAndReport(
+		[
+			[
+				'ga-user-timing-commercial-start',
+				function runTrackPerformance() {
+					EventTimer.get().mark('commercialStart');
+					EventTimer.get().mark('commercialBootStart');
+				},
+			],
+		],
+		tags,
+	);
+
+	// Stub the command queue
+	// @ts-expect-error -- it’s a stub, not the whole Googletag object
+	window.googletag = {
+		cmd: [],
+	};
+
+	try {
+		const promises = loadModules(modules);
+
+		await Promise.resolve(promises).then(recordCommercialMetrics);
+	} catch (error) {
+		// report async errors in bootCommercial to Sentry with the commercial feature tag
+		reportError(error, 'commercial', tags);
+	}
+};
+
+export { type Modules, bootCommercial };
